chore: fix ci actions

+++ src/Component/Console/Block/TextBlock.php
@@ -2,6 +2,15 @@
 
 declare(strict_types=1);
 
+/*
+ * This file is part of the Neutomic package.
+ *
+ * (c) Saif Eddin Gmati <azjezz@protonmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Neu\Component\Console\Block;
 
 use Neu\Component\Console\Output\OutputInterface;
@@ -10,6 +19,11 @@ use Neu\Component\Console\Terminal;
 use Psl\Str;
 use Psl\Vec;
 
+/**
+ * The {@see TextBlock} class is used to display a block of text in the output.
+ *
+ * @psalm-suppress MissingThrowsDocblock
+ */
 final readonly class TextBlock implements BlockInterface
 {
     private OutputInterface $output;
@@ -24,11 +38,11 @@ final readonly class TextBlock implements BlockInterface
      */
     public function display(string $message, Verbosity $verbosity = Verbosity::Normal): self
     {
-        $message = Str\wrap($message, (int)((Terminal::getWidth() / 3) * 2), cut: true);
+        $message = Str\wrap($message, (int)(((int)(Terminal::getWidth() / 3)) * 2), cut: true);
         $message = Str\join(
             Vec\map(
                 Str\split($message, OutputInterface::END_OF_LINE),
-                static fn(string $chunk) => '   ' . $chunk
+                static fn (string $chunk) => '   ' . $chunk
             ),
             OutputInterface::END_OF_LINE,
         );
